requirements.txt
Django >=5.1, <5.2
djangorestframework >=3.15
django-cors-headers >=4.4
python-decouple >= 3.8
mysqlclient >= 2.2.4
django-allauth >= 65.0.2
requests >= 2.32.3
PyJWT >= 2.9.0
cryptography >= 43.0.1
dj-rest-auth[with_social] >= 6.0
djangorestframework-simplejwt >= 5.3